requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
asyncio==3.4.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
paypalrestsdk==1.13.3
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.3.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
